refactoring json and xml
git-svn-id: https://svn.code.sf.net/p/sdmlib/code-0/trunk@385 2b52b8e8-071c-4d2f-a868-389f512d1cbb

+++ src/org/sdmlib/serialization/bytes/ByteEntity.java
@@ -40,33 +40,35 @@ import org.sdmlib.serialization.interfaces.JSIMEntity;
  */
 public class ByteEntity implements JSIMEntity, ByteItem{
 	/** The Constant BIT OF A BYTE. */
-	public final static int BITOFBYTE=8;
-	public final static int TYPBYTE=1;
-	
+	public final static int BITOFBYTE = 8;
+	public final static int TYPBYTE = 1;
+
 	/** The Byte Typ. */
 	protected byte typ;
-	
+
 	/** The values. */
 	protected byte[] values;
 	private boolean visible;
-	
+
 	/**
 	 * Instantiates a new byte entity.
 	 */
-	public ByteEntity(){
-		
+	public ByteEntity() {
+
 	}
-	
+
 	/**
 	 * Instantiates a new byte entity.
-	 *
-	 * @param typ the typ
-	 * @param value the value
+	 * 
+	 * @param typ
+	 *            the typ
+	 * @param value
+	 *            the value
 	 */
-	public ByteEntity(byte typ, byte[] value){
+	public ByteEntity(byte typ, byte[] value) {
 		this.setValue(typ, value);
 	}
-	
+
 	/*
 	 * @see de.uni.kassel.peermessage.BaseEntity#getNewArray()
 	 */
@@ -82,10 +84,10 @@ public class ByteEntity implements JSIMEntity, ByteItem{
 	public JSIMEntity getNewObject() {
 		return new ByteEntity();
 	}
-	
+
 	/**
 	 * Gets the value.
-	 *
+	 * 
 	 * @return the value
 	 */
 	public byte[] getValue() {
@@ -94,25 +96,28 @@ public class ByteEntity implements JSIMEntity, ByteItem{
 
 	/**
 	 * Sets the value.
-	 *
-	 * @param value the new value
+	 * 
+	 * @param value
+	 *            the new value
 	 */
 	public void setValue(byte typ, byte[] value) {
-		this.typ=typ;
+		this.typ = typ;
 		this.values = value;
 	}
 
 	/**
 	 * Byte to unsigned byte.
-	 *
-	 * @param n the Byte
+	 * 
+	 * @param n
+	 *            the Byte
 	 * @return the Byte
 	 */
 	public byte byteToUnsignedByte(int n) {
-		if (n < 128) return (byte) n;
+		if (n < 128)
+			return (byte) n;
 		return (byte) (n - 256);
 	}
-	
+
 	/*
 	 * @see de.uni.kassel.peermessage.Entity#toString()
 	 */
@@ -120,30 +125,38 @@ public class ByteEntity implements JSIMEntity, ByteItem{
 	public String toString() {
 		return toString(null);
 	}
-	public String toString(int indentFactor){
+
+	public String toString(int indentFactor) {
 		return toString(null);
 	}
 
 	/**
 	 * Convert the bytes to a String
-	 * @param converter Grammar
+	 * 
+	 * @param converter
+	 *            Grammar
 	 * @return converted bytes as String
 	 */
-	public String toString(ByteConverter converter){
+	public String toString(ByteConverter converter) {
 		return toString(converter, false);
 	}
+
 	/**
 	 * Convert the bytes to a String
-	 * @param converter Grammar
-	 * @param dynamic if byte is dynamic
+	 * 
+	 * @param converter
+	 *            Grammar
+	 * @param dynamic
+	 *            if byte is dynamic
 	 * @return converted bytes as String
 	 */
-	public String toString(ByteConverter converter, boolean dynamic){
-		if(converter==null){
-			converter=new ByteConverterHTTP();
+	public String toString(ByteConverter converter, boolean dynamic) {
+		if (converter == null) {
+			converter = new ByteConverterHTTP();
 		}
 		return converter.toString(this, dynamic);
 	}
+
 	/*
 	 * @see de.uni.kassel.peermessage.Entity#toString(int, int)
 	 */
@@ -151,114 +164,119 @@ public class ByteEntity implements JSIMEntity, ByteItem{
 	public String toString(int indentFactor, int intent) {
 		return toString(null);
 	}
-	
+
 	/**
 	 * Gets the bytes.
-	 *
+	 * 
 	 * @return the bytes
 	 */
-	public ByteBuffer getBytes(boolean isDynamic){
-		int len=calcLength(isDynamic);
-		byte typ=getTyp();
-		byte[] value=this.values;
-		
-		if(isDynamic&&value!=null){
+	public ByteBuffer getBytes(boolean isDynamic) {
+		int len = calcLength(isDynamic);
+		byte typ = getTyp();
+		byte[] value = this.values;
+
+		if (isDynamic && value != null) {
 			ByteBuffer bb = ByteBuffer.wrap(value);
-			if(typ==ByteIdMap.DATATYPE_SHORT){
-				short bufferValue=bb.getShort();
-				if(bufferValue>=Byte.MIN_VALUE&&bufferValue<=Byte.MAX_VALUE){
-					typ=ByteIdMap.DATATYPE_BYTE;
-					value=new byte[]{(byte)bufferValue};
+			if (typ == ByteIdMap.DATATYPE_SHORT) {
+				short bufferValue = bb.getShort();
+				if (bufferValue >= Byte.MIN_VALUE
+						&& bufferValue <= Byte.MAX_VALUE) {
+					typ = ByteIdMap.DATATYPE_BYTE;
+					value = new byte[] { (byte) bufferValue };
 				}
-			}else if(typ==ByteIdMap.DATATYPE_INTEGER||typ==ByteIdMap.DATATYPE_LONG){
-				int bufferValue=bb.getInt();
-				if(bufferValue>=Byte.MIN_VALUE&&bufferValue<=Byte.MAX_VALUE){
-					typ=ByteIdMap.DATATYPE_BYTE;
-					value=new byte[]{(byte)bufferValue};
-				}else if(bufferValue>=Short.MIN_VALUE&&bufferValue<=Short.MAX_VALUE){
-					typ=ByteIdMap.DATATYPE_BYTE;
-					ByteBuffer buffer = ByteBuffer.allocate(Short.SIZE/BITOFBYTE);
-					buffer.putShort((short)bufferValue);
+			} else if (typ == ByteIdMap.DATATYPE_INTEGER
+					|| typ == ByteIdMap.DATATYPE_LONG) {
+				int bufferValue = bb.getInt();
+				if (bufferValue >= Byte.MIN_VALUE
+						&& bufferValue <= Byte.MAX_VALUE) {
+					typ = ByteIdMap.DATATYPE_BYTE;
+					value = new byte[] { (byte) bufferValue };
+				} else if (bufferValue >= Short.MIN_VALUE
+						&& bufferValue <= Short.MAX_VALUE) {
+					typ = ByteIdMap.DATATYPE_BYTE;
+					ByteBuffer buffer = ByteBuffer.allocate(Short.SIZE
+							/ BITOFBYTE);
+					buffer.putShort((short) bufferValue);
 					buffer.flip();
-					value=buffer.array();
+					value = buffer.array();
 				}
 			}
 		}
 		ByteBuffer buffer = ByteUtil.getBuffer(len, typ);
-		
+
 		// Save the Len
-		if(value!=null){
+		if (value != null) {
 			buffer.put(value);
 		}
 		buffer.flip();
 		return buffer;
 	}
-	
-	public boolean setValues(Object value){
+
+	public boolean setValues(Object value) {
 		byte typ = 0;
 		ByteBuffer msgValue = null;
-		if(value ==null){
-			typ=ByteIdMap.DATATYPE_NULL;
+		if (value == null) {
+			typ = ByteIdMap.DATATYPE_NULL;
 		}
-		if(value instanceof Short){
-			typ=ByteIdMap.DATATYPE_SHORT;
-			msgValue = ByteBuffer.allocate(Short.SIZE/BITOFBYTE);
+		if (value instanceof Short) {
+			typ = ByteIdMap.DATATYPE_SHORT;
+			msgValue = ByteBuffer.allocate(Short.SIZE / BITOFBYTE);
 			msgValue.putShort((Short) value);
 		} else if (value instanceof Integer) {
-			typ=ByteIdMap.DATATYPE_INTEGER;
-			msgValue = ByteBuffer.allocate(Integer.SIZE/BITOFBYTE);
+			typ = ByteIdMap.DATATYPE_INTEGER;
+			msgValue = ByteBuffer.allocate(Integer.SIZE / BITOFBYTE);
 			msgValue.putInt((Integer) value);
 		} else if (value instanceof Long) {
-			typ=ByteIdMap.DATATYPE_LONG;
-			msgValue = ByteBuffer.allocate(Long.SIZE/BITOFBYTE);
+			typ = ByteIdMap.DATATYPE_LONG;
+			msgValue = ByteBuffer.allocate(Long.SIZE / BITOFBYTE);
 			msgValue.putLong((Long) value);
 		} else if (value instanceof Float) {
-			typ=ByteIdMap.DATATYPE_FLOAT;
-			msgValue = ByteBuffer.allocate(Float.SIZE/BITOFBYTE);
+			typ = ByteIdMap.DATATYPE_FLOAT;
+			msgValue = ByteBuffer.allocate(Float.SIZE / BITOFBYTE);
 			msgValue.putFloat((Float) value);
 		} else if (value instanceof Double) {
-			typ=ByteIdMap.DATATYPE_DOUBLE;
-			msgValue = ByteBuffer.allocate(Double.SIZE/BITOFBYTE);
-			msgValue.putDouble((Float) value);
+			typ = ByteIdMap.DATATYPE_DOUBLE;
+			msgValue = ByteBuffer.allocate(Double.SIZE / BITOFBYTE);
+			msgValue.putDouble((Double) value);
 		} else if (value instanceof Byte) {
-			typ=ByteIdMap.DATATYPE_BYTE;
-			msgValue = ByteBuffer.allocate(Byte.SIZE/BITOFBYTE);
+			typ = ByteIdMap.DATATYPE_BYTE;
+			msgValue = ByteBuffer.allocate(Byte.SIZE / BITOFBYTE);
 			msgValue.put((Byte) value);
 		} else if (value instanceof Character) {
-			typ=ByteIdMap.DATATYPE_CHAR;
-			msgValue = ByteBuffer.allocate(Character.SIZE/BITOFBYTE);
+			typ = ByteIdMap.DATATYPE_CHAR;
+			msgValue = ByteBuffer.allocate(Character.SIZE / BITOFBYTE);
 			msgValue.putChar((Character) value);
 		} else if (value instanceof String) {
-			typ=ByteIdMap.DATATYPE_STRING;
+			typ = ByteIdMap.DATATYPE_STRING;
 			String newValue = (String) value;
 			msgValue = ByteBuffer.allocate(newValue.length());
 			msgValue.put(newValue.getBytes());
 		} else if (value instanceof Date) {
-			typ=ByteIdMap.DATATYPE_DATE;
-			msgValue = ByteBuffer.allocate(Integer.SIZE/BITOFBYTE);
+			typ = ByteIdMap.DATATYPE_DATE;
+			msgValue = ByteBuffer.allocate(Integer.SIZE / BITOFBYTE);
 			Date newValue = (Date) value;
 			msgValue.putInt((int) newValue.getTime());
-		} else if (value instanceof Byte[]||value instanceof byte[]){
-			typ=ByteIdMap.DATATYPE_BYTEARRAY;
+		} else if (value instanceof Byte[] || value instanceof byte[]) {
+			typ = ByteIdMap.DATATYPE_BYTEARRAY;
 			byte[] newValue = (byte[]) value;
 			msgValue = ByteBuffer.allocate(newValue.length);
 			msgValue.put(newValue);
 		}
-		if(typ!=0){
-			this.typ=typ;
+		if (typ != 0) {
+			this.typ = typ;
 			// Check for group
-			if(msgValue!=null){
+			if (msgValue != null) {
 				msgValue.flip();
-				this.values=msgValue.array();
+				this.values = msgValue.array();
 			}
 			return true;
 		}
 		return false;
 	}
-	
+
 	/**
 	 * Gets the typ.
-	 *
+	 * 
 	 * @return the typ
 	 */
 	public byte getTyp() {
@@ -267,35 +285,39 @@ public class ByteEntity implements JSIMEntity, ByteItem{
 
 	/**
 	 * calculate the length of value
-	 *
+	 * 
 	 * @return the length
 	 */
 	public int calcLength(boolean isDynamic) {
 		// Length calculate Sonderfaelle ermitteln
-		if(isDynamic&&this.values!=null){
+		if (isDynamic && this.values != null) {
 			ByteBuffer bb = ByteBuffer.wrap(values);
-			if(typ==ByteIdMap.DATATYPE_SHORT){
-				Short bufferValue=bb.getShort();
-				if(bufferValue>=Byte.MIN_VALUE&&bufferValue<=Byte.MAX_VALUE){
-					return TYPBYTE+Byte.SIZE/BITOFBYTE;
+			if (typ == ByteIdMap.DATATYPE_SHORT) {
+				Short bufferValue = bb.getShort();
+				if (bufferValue >= Byte.MIN_VALUE
+						&& bufferValue <= Byte.MAX_VALUE) {
+					return TYPBYTE + Byte.SIZE / BITOFBYTE;
 				}
-			}else if(typ==ByteIdMap.DATATYPE_INTEGER||typ==ByteIdMap.DATATYPE_LONG){
-				Integer bufferValue=bb.getInt();
-				if(bufferValue>=Byte.MIN_VALUE&&bufferValue<=Byte.MAX_VALUE){
-					return TYPBYTE+Byte.SIZE/BITOFBYTE;
-				}else if(bufferValue>=Short.MIN_VALUE&&bufferValue<=Short.MAX_VALUE){
-					return TYPBYTE+Short.SIZE/BITOFBYTE;
+			} else if (typ == ByteIdMap.DATATYPE_INTEGER
+					|| typ == ByteIdMap.DATATYPE_LONG) {
+				Integer bufferValue = bb.getInt();
+				if (bufferValue >= Byte.MIN_VALUE
+						&& bufferValue <= Byte.MAX_VALUE) {
+					return TYPBYTE + Byte.SIZE / BITOFBYTE;
+				} else if (bufferValue >= Short.MIN_VALUE
+						&& bufferValue <= Short.MAX_VALUE) {
+					return TYPBYTE + Short.SIZE / BITOFBYTE;
 				}
 			}
 		}
-		int len=TYPBYTE+ByteUtil.getTypLen(getTyp());
-		
-		if(this.values!=null){
-			len+=this.values.length;
+		int len = TYPBYTE + ByteUtil.getTypLen(getTyp());
+
+		if (this.values != null) {
+			len += this.values.length;
 		}
 		return len;
 	}
-	
+
 	/**
 	 * Sets the len check.
 	 * 
@@ -305,8 +327,8 @@ public class ByteEntity implements JSIMEntity, ByteItem{
 	 */
 	public boolean setLenCheck(boolean isLenCheck) {
 		if (!isLenCheck) {
-			if(typ/16==(ByteIdMap.DATATYPE_CHECK/16)){
-			}else if(ByteUtil.isGroup(typ)){
+			if (typ / 16 == (ByteIdMap.DATATYPE_CHECK / 16)) {
+			} else if (ByteUtil.isGroup(typ)) {
 				this.typ = ByteUtil.getTyp(typ, ByteIdMap.DATATYPE_STRINGLAST);
 			}
 		} else {
@@ -317,7 +339,7 @@ public class ByteEntity implements JSIMEntity, ByteItem{
 	}
 
 	public void setVisible(boolean value) {
-		this.visible=value;
+		this.visible = value;
 	}
 
 	public boolean isVisible() {
